tnns/services/odds-ingestion/requirements.txt
kafka-python==2.0.2
requests==2.32.3
betfairlightweight==2.16.1
ciso8601==2.3.1
ujson==5.10.0
